Simplify Header component
Figured out bracket syntax on objects (it was working all along but
cssModules wasn’t providing an error so the undefined result was
unhelpful).

We now map an ordered array of strings representing the nav items which
we use to key off an object with the actual nav item objects and state.

Much less code. Not sure if this is the best pattern yet but boy howdy
is it better than what we had before :P