[#65] fix double loading bug
Items and monsters were both loading twice. This was due to a bug where
I was calling loadMonsters from both the monsters container and the
monstersList container. This is not necessary since I am passing the
monsters data from the monsters container down - which now that I think
about it probably means we have an extra unneeded container in there
that does nothing at all. Boo!
This bug was identical in both monsters and items. It is now fix
although I think further enhancements could be made.
—
Went ahead and removed the unneeded containers from monsters and items.

Makes me wonder what else is extra in here that can be stripped away…