History, log, note, reducers
Added history log and note reducers with tests. History uses log as a
child reducer and log uses note as a child reducer. This is done due to
the deep state setup with history being an array of logs that must
support CRUD and logs have an array of notes that must support CRUD.

- new action types
- history, log, note reducers
- unit tests