Campaign Log
Closes #93

Includes Note & Log components with test-component pages.

Log Test Container with state for full CRUD in test app. Useful for
seeing how all that the parts will fit together in the actual app.
There is quite a bit of function pass through. Plenty of comments in
the log test container that should help explain it.