login / signup toggle form
To avoid having to call ping from the login / logout actions, there is
now a login / signup form. This makes using an email for user
identification possible in an acceptable user flow. The email was
necessary for gravatars - ping was required for repainting the auth
header but now we handle that through the login action types and auth
reducer.