Clear app state on logout
Root reducer listens for log_out_success action type and sets state to
undefined before passing it of to the app reducer. Since app reducer is
just combine reducers it passes undefined down to all the reducers
resulting in each reducer returning it’s initial state, thus we reset
app state. If we do not do this a user can logout and then navigate to
routes with secure content that has been cached in state thus making
the logout appear unsuccessful when in fact it was.