async bugfix
bug was actually with mongoose / mongo. I changed the model after some
records had been added so they did not take the new model and were not
being returned by mongoose. I fixed this by explicitly stating which
collection the model is concerned with. also tests and removing the
redirect after selecting the campaign because it is buggy at the moment.