Clear logs in state when campaign loads
share an LOAD_CAMPAIGN_SUCCESS action type with the history reducer so
we can clear out the logs already in state. This clears out the logs
and forces the new and correct set to be to loaded in. Load if needed
checks for logs not if they are the ‘right’ ones.
Also improved the tests to check state change from accurate default
states instead of throwing everything through initial state.