Clear session cookie on logout
Clear the session cookie on logout. This clears out the campaign id in
cookie and forces campaign selection. We do not confirm that the
campaign id belongs to the logged in user. This handles that although
now I think we have a new bug in that a malicious user could
potentially read someone else’s campaign if they manipulated the cookie.