Bug fix
The fixes the unique key on iterators bug encountered when adding new
notes to a log.

This was caused because temp notes do not have an _id. That _id was
used for the key. Providing an _id with uuid() solved the problem but
caused notes to error when attempting to save to the db. Instead we
test for _id in the log component (where notes are iterated over) and
if it exists we use it otherwise we use a uuid without ever adding an
_id to the data itself.

Also in this commit - enzyme setup and a couple tests.